backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-multipart
httpx
openai
pypdf2
python-decouple
requests
aiofiles
PyMuPDF
chromadb
google-generativeai
